requirements.txt
Django==1.8.5
djangorestframework==3.1.1
elasticsearch==2.1.0
keyring==5.3
psycopg2==2.5.4
